+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-10 19:30:22
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-05-11 21:16:45
+// Commit: feat(perf): add bi-directional event bus handler for perf
 
 class perfComponentState {
-  final String commitRef = '1029';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-05-10T19:30:22+05:00');
+  final String commitRef = '1052';
+  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-11T21:16:45+05:00');
 }